agent/requirements.txt
livekit >= 0.18.0
livekit-protocol
livekit-agents>=0.11.0
livekit-plugins-openai>=0.10.5
fastapi>=0.100.0
uvicorn>=0.20.0
PyPDF2>=3.0.0
pdf2image>=1.16.0
pytesseract>=0.3.10
python-dotenv>=1.0.0
requests>=2.31.0
